refactor: adiciona contexto geral de layout de autenticação
